backend/requirements.txt
# FastAPI y dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Manejo de PDFs
PyMuPDF==1.23.8
reportlab==4.0.4
pikepdf==8.10.0
pdf2image==1.16.3
Pillow==10.1.0

# CSV y datos
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0

# Otros
python-dateutil==2.8.2
pydantic[email]==2.4.2
pydantic-settings==2.1.0